sessions state hardening
