+++ frontend/src/ui/status/runtimeOperationStatus.test.ts
@@ -3,8 +3,9 @@ import { describe, expect, it } from 'vitest';
 import {
   initialRuntimeOperationStatusState,
   runtimeOperationStatusReducer,
+  selectRuntimeOperationRows,
   type RuntimeOperationStatusState,
-} from './runtimeOperationStatus';
+} from './runtimeOperationStatusAdapter';
 
 const shellSession = {
   sessionId: 'shell-1',
@@ -111,5 +112,118 @@ describe('runtimeOperationStatusReducer', () => {
       localPort: 18081,
       podName: 'web-replacement',
     });
+  });
+
+  it('does not resurrect missing port-forward details from status-only events', () => {
+    const state: RuntimeOperationStatusState = {
+      operationsLoaded: true,
+      operations: [
+        {
+          id: 'pf-1',
+          type: 'port-forward',
+          clusterId: 'cluster-a',
+          status: 'active',
+          startedAt: '2026-05-18T00:00:00Z',
+        },
+      ],
+      shellSessions: [],
+      portForwardSessions: [],
+    };
+
+    const next = runtimeOperationStatusReducer(state, {
+      type: 'portforward:status',
+      event: {
+        sessionId: 'pf-1',
+        status: 'active',
+        localPort: 18080,
+        podName: 'web',
+      },
+    });
+
+    expect(next.portForwardSessions).toEqual([]);
+  });
+});
+
+describe('selectRuntimeOperationRows', () => {
+  it('shows workflow detail rows before the runtime operation registry loads', () => {
+    const state: RuntimeOperationStatusState = {
+      operationsLoaded: false,
+      operations: [],
+      shellSessions: [
+        { ...shellSession, sessionId: 'older', startedAt: '2026-05-18T00:00:00Z' },
+        { ...shellSession, sessionId: 'newer', startedAt: '2026-05-19T00:00:00Z' },
+      ],
+      portForwardSessions: [
+        { ...portForwardSession, id: 'pf-error', status: 'error' },
+        { ...portForwardSession, id: 'pf-active', status: 'active' },
+      ],
+    };
+
+    const rows = selectRuntimeOperationRows(state);
+
+    expect(rows.shellSessions.map((session) => session.sessionId)).toEqual(['newer', 'older']);
+    expect(rows.portForwardSessions.map((session) => session.id)).toEqual([
+      'pf-active',
+      'pf-error',
+    ]);
+  });
+
+  it('filters pre-load workflow detail rows by selected cluster', () => {
+    const state: RuntimeOperationStatusState = {
+      operationsLoaded: false,
+      operations: [],
+      shellSessions: [
+        shellSession,
+        { ...shellSession, sessionId: 'shell-b', clusterId: 'cluster-b' },
+      ],
+      portForwardSessions: [
+        portForwardSession,
+        { ...portForwardSession, id: 'pf-b', clusterId: 'cluster-b' },
+      ],
+    };
+
+    const rows = selectRuntimeOperationRows(state, 'cluster-a');
+
+    expect(rows.shellSessions.map((session) => session.sessionId)).toEqual(['shell-1']);
+    expect(rows.portForwardSessions.map((session) => session.id)).toEqual(['pf-1']);
+  });
+
+  it('uses selected-cluster runtime operations as the presence envelope after load', () => {
+    const state: RuntimeOperationStatusState = {
+      operationsLoaded: true,
+      operations: [
+        {
+          id: 'shell-1',
+          type: 'shell',
+          clusterId: 'cluster-a',
+          status: 'open',
+          startedAt: '2026-05-18T00:00:00Z',
+        },
+        {
+          id: 'shell-b',
+          type: 'shell',
+          clusterId: 'cluster-b',
+          status: 'open',
+          startedAt: '2026-05-18T00:00:00Z',
+        },
+        {
+          id: 'drain-1',
+          type: 'drain',
+          clusterId: 'cluster-a',
+          status: 'running',
+          startedAt: '2026-05-18T00:00:00Z',
+        },
+      ],
+      shellSessions: [
+        shellSession,
+        { ...shellSession, sessionId: 'shell-b', clusterId: 'cluster-b' },
+        { ...shellSession, sessionId: 'ghost-shell', clusterId: 'cluster-a' },
+      ],
+      portForwardSessions: [],
+    };
+
+    const rows = selectRuntimeOperationRows(state, 'cluster-a');
+
+    expect(rows.shellSessions.map((session) => session.sessionId)).toEqual(['shell-1']);
   });
 });
